refactor(boundaries): break cockpit <-> cockpit_rail cycle via metrics-rail extract (refs #1367)
The cockpit_rail module had to lazy-import _register_metrics_rail_item and
_register_worker_roster_rail_item from cockpit because cockpit top-imports
cockpit_rail for the router re-export. Moved the metrics-rail registration
into a new pollypm.cockpit_metrics_rail leaf module and sourced the worker
roster registration from its real home in cockpit_inbox, so cockpit_rail
can import both at module load time. cockpit.py keeps a re-export shim
for legacy callers / monkeypatch paths.
